Separate main class from file Rename some files
