+++ LA.sh
@@ -20,10 +20,10 @@ ARG2=$2
 # switch file and hwconfig argument order
 if [ "$ARG2" == "-s3d" ]; then
   echo "start script"
-  python3 $DIR/LAsimplify3D.py "${@}"
+  python3 $DIR/linearAdvance.py "${@}"
 elif [ "$ARG1" == "-ps" ]; then
   echo "start script"
-  python3 $DIR/LAsimplify3D.py "$2" "$1"
+  python3 $DIR/linearAdvance.py "$2" "$1"
 else
   echo "Error in argument of type of slicer"
 fi
